Refactor to use a hook so that we can get the blockNames via getBlockTypes and use useMemo Change function signature to something like getGlobalStylesChangelist( next, previous, options = {} ) Move max-length functionality from the revision buttons to the util.
