+++ packages/block-editor/src/components/global-styles/index.js
@@ -29,4 +29,4 @@ export {
 } from './image-settings-panel';
 export { default as AdvancedPanel } from './advanced-panel';
 export { areGlobalStyleConfigsEqual } from './utils';
-export { default as getGlobalStylesChangelist } from './get-global-styles-changelist';
+export { default as useGlobalStylesChangelist } from './use-global-styles-changelist';
